+++ src/scriptlets/prevent-addEventListener.js
@@ -4,6 +4,7 @@ import {
     validateType,
     validateListener,
     listenerToString,
+    logMessage,
 } from '../helpers';
 
 /* eslint-disable max-len */
@@ -21,14 +22,24 @@ import {
  *
  * ### Syntax
  *
+ * <!-- markdownlint-disable line-length -->
+ *
  * ```text
- * example.org#%#//scriptlet('prevent-addEventListener'[, typeSearch[, listenerSearch]])
+ * example.org#%#//scriptlet('prevent-addEventListener'[, typeSearch[, listenerSearch[, additionalArgName, additionalArgValue]]])
  * ```
+ *
+ * <!-- markdownlint-enable line-length -->
  *
  * - `typeSearch` — optional, string or regular expression matching the type (event name);
  *   defaults to match all types; invalid regular expression will cause exit and rule will not work
  * - `listenerSearch` — optional, string or regular expression matching the listener function body;
  *   defaults to match all listeners; invalid regular expression will cause exit and rule will not work
+ * - `additionalArgName` — optional, string, name of the additional argument to match;
+ *   currently only `elements` is supported;
+ * - `additionalArgValue` — optional, value corresponding to the additional argument name;
+ *   for `elements` it can be a CSS selector or one of the following values:
+ *     - `window`
+ *     - `document`
  *
  * ### Examples
  *
@@ -52,12 +63,75 @@ import {
  *     });
  *     ```
  *
+ * 1. Prevent 'click' listeners with the callback body containing `foo` and only if the element has the class `bar`
+ *
+ *     ```adblock
+ *     example.org#%#//scriptlet('prevent-addEventListener', 'click', 'foo', 'elements', '.bar')
+ *     ```
+ *
+ *     For instance, this listener will not be called:
+ *
+ *     ```javascript
+ *     const el = document.querySelector('.bar');
+ *     el.addEventListener('click', () => {
+ *         window.test = 'foo';
+ *     });
+ *     ```
+ *
+ *     This listener will be called:
+ *
+ *     ```javascript
+ *     const el = document.querySelector('.xyz');
+ *     el.addEventListener('click', () => {
+ *         window.test = 'foo';
+ *     });
+ *     ```
+ *
  * @added v1.0.4.
  */
 /* eslint-enable max-len */
-export function preventAddEventListener(source, typeSearch, listenerSearch) {
+export function preventAddEventListener(source, typeSearch, listenerSearch, additionalArgName, additionalArgValue) {
     const typeSearchRegexp = toRegExp(typeSearch);
     const listenerSearchRegexp = toRegExp(listenerSearch);
+
+    let elementToMatch;
+    if (additionalArgName) {
+        if (additionalArgName !== 'elements') {
+            logMessage(source, `Invalid "additionalArgName": ${additionalArgName}\nOnly "elements" is supported.`);
+            return;
+        }
+
+        if (!additionalArgValue) {
+            logMessage(source, '"additionalArgValue" is required.');
+            return;
+        }
+
+        elementToMatch = additionalArgValue;
+    }
+
+    /**
+     * Checks if an element matches the specified selector or element type.
+     *
+     * @param {any} element - The element to check
+     * @returns {boolean}
+     */
+    const elementMatches = (element) => {
+        // If elementToMatch is undefined, it means that the scriptlet was called without the `elements` argument
+        // so it should match all elements
+        if (elementToMatch === undefined) {
+            return true;
+        }
+        if (elementToMatch === 'window') {
+            return element === window;
+        }
+        if (elementToMatch === 'document') {
+            return element === document;
+        }
+        if (element && element.matches && element.matches(elementToMatch)) {
+            return true;
+        }
+        return false;
+    };
 
     const nativeAddEventListener = window.EventTarget.prototype.addEventListener;
 
@@ -65,7 +139,8 @@ export function preventAddEventListener(source, typeSearch, listenerSearch) {
         let shouldPrevent = false;
         if (validateType(type) && validateListener(listener)) {
             shouldPrevent = typeSearchRegexp.test(type.toString())
-                && listenerSearchRegexp.test(listenerToString(listener));
+                && listenerSearchRegexp.test(listenerToString(listener))
+                && elementMatches(this);
         }
 
         if (shouldPrevent) {
@@ -115,4 +190,5 @@ preventAddEventListener.injections = [
     validateType,
     validateListener,
     listenerToString,
+    logMessage,
 ];